requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
alembic==1.16.1
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
argon2-cffi==25.1.0
    # via passlib
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
certifi==2025.4.26
    # via
    #   httpcore
    #   httpx
cffi==1.17.1
    # via argon2-cffi-bindings
click==8.2.1
    # via
    #   rich-toolkit
    #   typer
    #   uvicorn
colorama==0.4.6
    # via
    #   click
    #   uvicorn
dnspython==2.7.0
    # via
    #   email-validator
    #   pymongo
ecdsa==0.19.1
    # via python-jose
email-validator==2.2.0
    # via fastapi
fastapi[standard]==0.115.12
    # via -r requirements.in
fastapi-cli[standard]==0.0.7
    # via fastapi
greenlet==3.2.2
    # via sqlalchemy
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via fastapi
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
jinja2==3.1.6
    # via fastapi
mako==1.3.10
    # via alembic
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   jinja2
    #   mako
mdurl==0.1.2
    # via markdown-it-py
passlib[argon2]==1.7.4
    # via -r requirements.in
pyasn1==0.6.1
    # via
    #   python-jose
    #   rsa
pycparser==2.22
    # via cffi
pydantic==2.11.5
    # via fastapi
pydantic-core==2.33.2
    # via pydantic
pygments==2.19.1
    # via rich
pymongo==4.13.0
    # via -r requirements.in
python-dotenv==1.1.0
    # via uvicorn
python-jose==3.5.0
    # via -r requirements.in
python-multipart==0.0.20
    # via fastapi
pyyaml==6.0.2
    # via uvicorn
rich==14.0.0
    # via
    #   rich-toolkit
    #   typer
rich-toolkit==0.14.7
    # via fastapi-cli
rsa==4.9.1
    # via python-jose
shellingham==1.5.4
    # via typer
six==1.17.0
    # via ecdsa
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.41
    # via
    #   -r requirements.in
    #   alembic
starlette==0.46.2
    # via fastapi
typer==0.16.0
    # via fastapi-cli
typing-extensions==4.14.0
    # via
    #   alembic
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   rich-toolkit
    #   sqlalchemy
    #   typer
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
uvicorn[standard]==0.34.3
    # via
    #   fastapi
    #   fastapi-cli
watchfiles==1.0.5
    # via uvicorn
websockets==15.0.1
    # via uvicorn
